garage-controller/requirements.txt
fastapi==0.110.0  # Web framework for the controller (adjust version as needed)
uvicorn==0.29.0   # ASGI server to run FastAPI
PyYAML==6.0.1     # YAML config parsing
requests==2.31.0  # Optional: may be useful for future internal calls/logging
